refactor: featureToggle and utils
